Fixes #8800: Move ScrollbarVisibility to /common/
